Implemented simple git clone step
As a step developer we often has to clone external repositories as
dependencies usually for testing purposes. For this the standard git
clone step is not usable since it uses hard coded bitrise input variables
also sets multiple key environment variables.

The usage is pretty similar to the standard clone step but is a bit more
relaxed and has less options.

[ACT-390]